Initial commit of the VisualDisplay component
The component now defines a memory region and listens to changes in
memory. It includes a 256-colors pallette that allows to draw colors in
the display by writing to the corresponding memory cells.

The CPUService has also been modified. Now the registers can be accessed
seamlessly and the code of the instructions can be defined more easily.