Improved CPU speed
Now the simulator is capable of executing at greater speeds.

In order to cope with the delay introduced by the Javascript and Angular
frameworks, the simulator now introduces a sort of *pipeline*. At
greater speeds, the simulator executes more than one instruction on each
timer interrupt. This has an undesirable effect, since the simulator
will not be capable of responding to any interrupt until all the
instructions of the cycle are executed.