Fixed marking of stack cells
Now the cells that actually belong to the stack are correctly marked on
the memory view. The SP registers are now treated as simple registers.
The memory view controller has been simplified. The mapping of the cells
that belong to the stack is now done depending on the operations
executed over the stack pointer.